py-server/functions/requirements.lock.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
base58==2.1.1
bitarray==3.4.1
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==1.0.2
click==8.2.0
cloudevents==1.9.0
construct==2.10.68
construct-typing==0.5.6
cryptography==45.0.2
cytoolz==1.0.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
firebase-admin==6.5.0
firebase-functions==0.1.2
Flask==3.1.1
flask-cors==6.0.0
frozenlist==1.6.0
functions-framework==3.8.3
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-pubsub==2.29.1
google-cloud-storage==3.1.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-events==0.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonalias==0.1.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lru-dict==1.2.0
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.4.4
numpy==2.2.6
openai==1.79.0
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-propagator-gcp==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pandas==2.2.3
parsimonious==0.10.0
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
py-ecc==8.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.25.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
solana==0.33.0
solders==0.21.0
tiktoken==0.9.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
web3==6.15.1
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
